Disabled Buttons
When you click on a field, sometimes it shows the Values and Distributions
buttons, but sometimes it doesn't. It can be confusing what the criteria
are for them to show up, so instead of just randomly disappearing them, have
them show up disabled with an explanation for why they aren't working.